+++ .github/CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this package will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/), and this project *somewhat* adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.6.0] - 2020-07-10
+### Added
+ - Added the following class and method :
+   - Class : **ScriptingTools.Math**
+     - `RoundToStep(float, float)`
+     - `RoundToStepInt(float, int)`
+
 ## [0.5.0] - 2020-06-09
 ### Added
  - Added the following class and method :